shifted require_super_admin to acl_decorators as new decorator.
